rewrited mobs sync
