+++ GameDataSync/GameDataSync.LevelGraphCapture.cs
@@ -1,728 +1,21 @@
-using DeadCellsMultiplayerMod.Interface.ModuleInitializing;
-using ModCore.Events;
 using dc;
 using dc.haxe.ds;
 using dc.level;
-using dc.pr;
-using dc.tool;
 using Hashlink.Virtuals;
 using HaxeProxy.Runtime;
 using ModCore.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.Runtime.InteropServices;
 using System.Reflection;
-using System.Text;
-using System.Text.Json;
-using System.Threading;
 using dc.haxe;
-using dc.haxe.io;
 using dc.hl.types;
-using Rand = dc.libs.Rand;
 
 
 namespace DeadCellsMultiplayerMod
 {
     internal partial class GameDataSync
     {
-
-        private static void CopyStoryVisitedLoreRoomsToSet(dynamic? map, HashSet<string> target)
-        {
-            target.Clear();
-            if (map == null)
-                return;
-
-            try
-            {
-                var keys = map.keys.Invoke();
-                while (keys.hasNext.Invoke())
-                {
-                    var keyObj = keys.next.Invoke();
-                    if (keyObj == null)
-                        continue;
-
-                    var key = keyObj.ToString();
-                    if (string.IsNullOrWhiteSpace(key))
-                        continue;
-
-                    var raw = map.get.Invoke(keyObj);
-                    var visited = raw is bool b ? b : ToInt(raw) != 0;
-                    if (visited)
-                        target.Add(key);
-                }
-            }
-            catch
-            {
-            }
-        }
-
-        private static void CopyStoryPlannedLoresToList(ArrayBytes_Int? source, List<int> target)
-        {
-            target.Clear();
-            if (source == null)
-                return;
-
-            var seen = new HashSet<int>();
-            for (var i = 0; i < source.length; i++)
-            {
-                int planned;
-                try
-                {
-                    planned = ToInt(source.getDyn(i));
-                }
-                catch
-                {
-                    continue;
-                }
-
-                if (seen.Add(planned))
-                    target.Add(planned);
-            }
-        }
-
-        private static StringMap BuildCountersMap(Dictionary<string, int> values)
-        {
-            var map = new StringMap();
-            foreach (var kv in values)
-                map.set(kv.Key.AsHaxeString(), kv.Value);
-            return map;
-        }
-
-        private static EnumValueMap BuildNpcProgressMap(Dictionary<int, int> values)
-        {
-            var map = new EnumValueMap();
-            foreach (var kv in values)
-            {
-                var npcId = CreateNpcIdFromIndex(kv.Key);
-                if (npcId == null)
-                    continue;
-
-                map.set(npcId, kv.Value);
-            }
-            return map;
-        }
-
-        private static void ApplyStoryStringIntMap(dynamic? map, Dictionary<string, int> values)
-        {
-            if (map == null)
-                return;
-
-            try
-            {
-                var keysToRemove = new List<object>();
-                var keys = map.keys.Invoke();
-                while (keys.hasNext.Invoke())
-                {
-                    var keyObj = keys.next.Invoke();
-                    if (keyObj != null)
-                        keysToRemove.Add(keyObj);
-                }
-
-                for (var i = 0; i < keysToRemove.Count; i++)
-                {
-                    try { map.remove.Invoke(keysToRemove[i]); } catch { }
-                }
-            }
-            catch
-            {
-            }
-
-            foreach (var kv in values)
-            {
-                try
-                {
-                    map.set.Invoke(kv.Key.AsHaxeString(), kv.Value);
-                }
-                catch
-                {
-                }
-            }
-        }
-
-        private static void ApplyStoryVisitedLoreRoomsMap(dynamic? map, HashSet<string> values)
-        {
-            if (map == null)
-                return;
-
-            try
-            {
-                var keysToRemove = new List<object>();
-                var keys = map.keys.Invoke();
-                while (keys.hasNext.Invoke())
-                {
-                    var keyObj = keys.next.Invoke();
-                    if (keyObj != null)
-                        keysToRemove.Add(keyObj);
-                }
-
-                for (var i = 0; i < keysToRemove.Count; i++)
-                {
-                    try { map.remove.Invoke(keysToRemove[i]); } catch { }
-                }
-            }
-            catch
-            {
-            }
-
-            foreach (var key in values)
-            {
-                try
-                {
-                    map.set.Invoke(key.AsHaxeString(), 1);
-                }
-                catch
-                {
-                }
-            }
-        }
-
-        private static ArrayBytes_Int BuildStoryPlannedLoresArray(List<int> values)
-        {
-            var arr = new ArrayBytes_Int();
-            var seen = new HashSet<int>();
-            for (var i = 0; i < values.Count; i++)
-            {
-                var planned = values[i];
-                if (!seen.Add(planned))
-                    continue;
-
-                try
-                {
-                    arr.push(planned);
-                }
-                catch
-                {
-                    try
-                    {
-                        arr.pushDyn(planned);
-                    }
-                    catch
-                    {
-                    }
-                }
-            }
-
-            return arr;
-        }
-
-        private static void ApplyStoryState(
-            User user,
-            Dictionary<string, int> counters,
-            Dictionary<int, int> npcProgress,
-            int storyDataVersion,
-            Dictionary<string, int> loreRoomRunIds,
-            HashSet<string> visitedLoreRooms,
-            List<int> plannedLores)
-        {
-            user.story = BuildStoryManager(
-                counters,
-                npcProgress,
-                storyDataVersion,
-                loreRoomRunIds,
-                visitedLoreRooms,
-                plannedLores);
-            user.counters = BuildCountersMap(counters);
-            user.npcs = BuildNpcProgressMap(npcProgress);
-        }
-
-        private static void ClearUserStoryState(User user)
-        {
-            user.story = null;
-            user.counters = new StringMap();
-            user.npcs = new EnumValueMap();
-        }
-
-        private static StoryManager BuildStoryManager(
-            Dictionary<string, int> counters,
-            Dictionary<int, int> npcProgress,
-            int storyDataVersion,
-            Dictionary<string, int> loreRoomRunIds,
-            HashSet<string> visitedLoreRooms,
-            List<int> plannedLores)
-        {
-            var story = new StoryManager();
-            try
-            {
-                story.onReload();
-            }
-            catch
-            {
-            }
-
-            dynamic dynStory = story;
-            dynStory.counters = BuildCountersMap(counters);
-            dynStory.npcProgresses = BuildNpcProgressMap(npcProgress);
-            ApplyStoryStringIntMap(dynStory.loreRoomRunIds, loreRoomRunIds);
-            ApplyStoryVisitedLoreRoomsMap(dynStory.visitedLoreRooms, visitedLoreRooms);
-            dynStory.plannedLores = BuildStoryPlannedLoresArray(plannedLores);
-            dynStory.storyDataVersion = storyDataVersion;
-            return story;
-        }
-
-        private static ArrayObj? CloneItemProgress(ArrayObj? source)
-        {
-            if (source == null)
-                return null;
-
-            var arr = ArrayUtils.CreateDyn();
-            for (int i = 0; i < source.length; i++)
-            {
-                var item = source.getDyn(i) as ItemProgress;
-                if (item == null)
-                    continue;
-                var copy = new ItemProgress(item.itemId);
-                copy.investedCells = item.investedCells;
-                copy.isNew = item.isNew;
-                copy.unlocked = item.unlocked;
-                copy.__uid = item.__uid;
-                arr.array.pushDyn(copy);
-            }
-            return (ArrayObj)arr.array;
-        }
-
-        private static ArrayObj? CloneMetaProgress(ArrayObj? source)
-        {
-            if (source == null)
-                return null;
-
-            var arr = ArrayUtils.CreateDyn();
-            for (int i = 0; i < source.length; i++)
-            {
-                var item = source.getDyn(i) as MetaProgress;
-                if (item == null)
-                    continue;
-
-                var copy = new MetaProgress(item.itemId);
-                copy.investedCells = item.investedCells;
-                copy.isNew = item.isNew;
-                copy.unlocked = item.unlocked;
-                copy.upgradeLevel = item.upgradeLevel;
-                copy.n = item.n;
-                copy.done = item.done;
-                copy.metaLevel = item.metaLevel;
-                arr.array.pushDyn(copy);
-            }
-
-            return (ArrayObj)arr.array;
-        }
-
-        private static ArrayObj? CloneItemList(ArrayObj? source)
-        {
-            if (source == null)
-                return null;
-
-            var arr = ArrayUtils.CreateDyn();
-            for (int i = 0; i < source.length; i++)
-            {
-                object? item = source.getDyn(i);
-                arr.array.pushDyn(item);
-            }
-            return (ArrayObj)arr.array;
-        }
-
-        private static IntMap? CloneIntMap(IntMap? source)
-        {
-            if (source == null)
-                return null;
-
-            var map = new IntMap();
-            try
-            {
-                var keys = source.keys();
-                while (keys.hasNext.Invoke())
-                {
-                    var key = keys.next.Invoke();
-                    map.set(key, ToInt(source.get(key)));
-                }
-            }
-            catch
-            {
-            }
-
-            return map;
-        }
-
-        internal static bool TryBuildSafeSaveUser(User currentUser, bool onlyGameData, out User? saveUser)
-        {
-            saveUser = null;
-            if (currentUser == null || string.IsNullOrWhiteSpace(_origProgressPayload))
-                return false;
-
-            try
-            {
-                var cloneBytes = Save.Class.genSave.Invoke(currentUser, onlyGameData);
-                if (cloneBytes == null)
-                    return false;
-
-                var clone = Save.Class.readSave.Invoke(cloneBytes);
-                if (clone == null)
-                    return false;
-
-                if (!TryApplyProgressPayload(clone, _origProgressPayload))
-                    return false;
-
-                saveUser = clone;
-                return true;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
-        private static bool TryApplyProgressPayload(User target, string? payload)
-        {
-            if (target == null || string.IsNullOrWhiteSpace(payload))
-                return false;
-
-            if (!TryDeserializeProgressUser(payload, out var source) || source == null)
-                return false;
-
-            ApplyProgressSnapshot(target, source);
-            return true;
-        }
-
-        private static void ApplyProgressSnapshot(User target, User source)
-        {
-            if (target == null || source == null)
-                return;
-
-            var preservedHeroSkin = !string.IsNullOrWhiteSpace(_origHeroSkin)
-                ? _origHeroSkin
-                : CleanSkin(target.heroSkin?.ToString());
-            var preservedHeroHeadSkin = !string.IsNullOrWhiteSpace(_origHeroHeadSkin)
-                ? _origHeroHeadSkin
-                : CleanSkin(target.heroHeadSkin?.ToString());
-
-            var legacyCounters = new Dictionary<string, int>(StringComparer.Ordinal);
-            CopyCountersToDictionary(source.counters, legacyCounters);
-            var legacyNpcProgress = new Dictionary<int, int>();
-            CopyNpcProgressToDictionary(source.npcs, legacyNpcProgress);
-            var sourceStory = source.story;
-            if (sourceStory != null)
-            {
-                if (sourceStory.counters == null)
-                    sourceStory.counters = BuildCountersMap(legacyCounters);
-                if (sourceStory.npcProgresses == null)
-                    sourceStory.npcProgresses = BuildNpcProgressMap(legacyNpcProgress);
-            }
-
-            target.flags = source.flags;
-            target.userId = source.userId;
-            target.deathMoney = source.deathMoney;
-            target.deathCells = source.deathCells;
-            target.bossRuneActivated = GetEffectiveBossRune(source);
-            target.tutorial = source.tutorial;
-            target.counters = sourceStory?.counters ?? BuildCountersMap(legacyCounters);
-            target.npcs = sourceStory?.npcProgresses ?? BuildNpcProgressMap(legacyNpcProgress);
-            target.story = sourceStory;
-            target.itemMeta = null;
-            target.userStats = source.userStats;
-            target.activeMods = CloneItemList(source.activeMods);
-            target.meta = CloneMetaProgress(source.meta);
-            target.metaItems = CloneItemList(source.metaItems);
-            target.achievements = CloneItemList(source.achievements);
-            target.localAchievements = CloneItemList(source.localAchievements);
-            target.deathItem = source.deathItem;
-            target.consecutiveCompletedRuns = source.consecutiveCompletedRuns;
-            ApplyHeroCosmetics(target, preservedHeroSkin, preservedHeroHeadSkin);
-            MirrorStoryStateToSaveUser(target);
-
-            try
-            {
-                target.userStats?.init();
-            }
-            catch
-            {
-            }
-
-            try
-            {
-                target.onReload();
-            }
-            catch
-            {
-            }
-
-            ApplyHeroCosmetics(target, preservedHeroSkin, preservedHeroHeadSkin);
-            MirrorStoryStateToSaveUser(target);
-
-            try
-            {
-                target.story?.onReload();
-            }
-            catch
-            {
-            }
-
-            var targetMeta = EnsureItemMeta(target, target.itemMeta);
-            targetMeta._user = target;
-            if (source.itemMeta != null)
-            {
-                targetMeta.itemProgress = CloneItemProgress(source.itemMeta.itemProgress) ?? EnsureArray(targetMeta.itemProgress);
-                targetMeta.permanentItems = CloneItemList(source.itemMeta.permanentItems) ?? EnsureArray(targetMeta.permanentItems);
-                targetMeta.forgeInvestedCells = CloneIntMap(source.itemMeta.forgeInvestedCells) ?? new IntMap();
-            }
-
-            try
-            {
-                targetMeta.onReload();
-            }
-            catch
-            {
-            }
-
-            try
-            {
-                targetMeta.revealAllBaseItems();
-            }
-            catch
-            {
-            }
-
-            try
-            {
-                targetMeta.cleanDuplicatedItemProgress();
-            }
-            catch
-            {
-            }
-
-            target.itemMeta = targetMeta;
-
-            try
-            {
-                target.br_setActivated(GetEffectiveBossRune(source));
-            }
-            catch
-            {
-                target.bossRuneActivated = GetEffectiveBossRune(source);
-            }
-        }
-
-        private static void MirrorStoryStateToSaveUser(User user)
-        {
-            if (user == null)
-                return;
-
-            try
-            {
-                var saveUser = user.mainGameData?.sUser;
-                if (saveUser == null || ReferenceEquals(saveUser, user))
-                    return;
-
-                saveUser.story = user.story;
-                saveUser.counters = user.counters;
-                saveUser.npcs = user.npcs;
-            }
-            catch
-            {
-            }
-        }
-
-        private static void ApplyHeroCosmetics(User? user, string? heroSkin, string? heroHeadSkin)
-        {
-            if (user == null)
-                return;
-
-            if (!string.IsNullOrWhiteSpace(heroSkin))
-                user.heroSkin = heroSkin.AsHaxeString();
-
-            if (!string.IsNullOrWhiteSpace(heroHeadSkin))
-                user.heroHeadSkin = heroHeadSkin.AsHaxeString();
-        }
-
-        private static string? BuildProgressPayload(User user)
-        {
-            if (user == null)
-                return null;
-
-            try
-            {
-                var prepared = false;
-                try
-                {
-                    prepared = user.prepareSave();
-                }
-                catch (Exception ex)
-                {
-                    _log?.Debug(ex, "[NetMod] user.prepareSave() threw while building progress payload");
-                }
-
-                if (!prepared)
-                    _log?.Debug("[NetMod] user.prepareSave() returned false; trying packed progress payload anyway");
-
-                var saveBytes = Save.Class.genSave.Invoke(user, true);
-                if (saveBytes != null)
-                {
-                    var saveRaw = CopyBytesToManaged(saveBytes);
-                    return saveRaw.Length == 0 ? "P2|" : "P2|" + Convert.ToBase64String(saveRaw);
-                }
-
-                _log?.Warning("[NetMod] Failed to build packed progress payload: save bytes were null");
-            }
-            catch (Exception ex)
-            {
-                _log?.Warning(ex, "[NetMod] Failed to build packed progress payload");
-            }
-
-            if (TrySerializeUserBytes(user, out var userBytes) && userBytes != null)
-            {
-                var userRaw = CopyBytesToManaged(userBytes);
-                return userRaw.Length == 0 ? "P1|" : "P1|" + Convert.ToBase64String(userRaw);
-            }
-
-            _log?.Warning("[NetMod] Failed to build progress payload from both packed-save and raw-user paths");
-            return null;
-        }
-
-        private static bool TryDeserializeProgressUser(string payload, out User? user)
-        {
-            user = null;
-            if (string.IsNullOrWhiteSpace(payload))
-                return false;
-
-            var isPackedSavePayload = payload.StartsWith("P2|", StringComparison.Ordinal);
-            var encoded =
-                payload.StartsWith("P2|", StringComparison.Ordinal) ? payload[3..] :
-                payload.StartsWith("P1|", StringComparison.Ordinal) ? payload[3..] :
-                payload;
-            byte[] raw;
-            try
-            {
-                raw = string.IsNullOrEmpty(encoded) ? Array.Empty<byte>() : Convert.FromBase64String(encoded);
-            }
-            catch
-            {
-                return false;
-            }
-
-            if (!TryCreateHaxeBytes(raw, out var bytes) || bytes == null)
-                return false;
-
-            try
-            {
-                if (isPackedSavePayload)
-                {
-                    user = Save.Class.readSave.Invoke(bytes);
-                    return user != null;
-                }
-
-                var serializer = new dc.hxbit.Serializer
-                {
-                    refs = new IntMap()
-                };
-                var position = 0;
-                serializer.beginLoad(bytes, Ref<int>.From(ref position));
-                user = (User)(object)serializer.getRef(User.Class, User.Class.__clid);
-                serializer.endLoad();
-                user?.onReload();
-                return user != null;
-            }
-            catch (Exception ex)
-            {
-                _log?.Warning(ex, "[NetMod] Failed to deserialize progress payload");
-                user = null;
-                return false;
-            }
-        }
-
-        private static bool TrySerializeUserBytes(User user, out Bytes? bytes)
-        {
-            bytes = null;
-            if (user == null)
-                return false;
-
-            try
-            {
-                var serializer = new dc.hxbit.Serializer();
-                serializer.beginSave();
-                var userRef = user.unnamedField0;
-                if (userRef == null)
-                    userRef = user.unnamedField0 = (virtual___uid_getCLID_getSerializeSchema_serialize_unserialize_unserializeInit_)(object)user;
-                serializer.addKnownRef(userRef);
-                var position = 0;
-                bytes = serializer.endSave(Ref<int>.From(ref position));
-                return bytes != null;
-            }
-            catch (Exception ex)
-            {
-                _log?.Warning(ex, "[NetMod] Failed to serialize raw user progress payload");
-                bytes = null;
-                return false;
-            }
-        }
-
-        private static byte[] CopyBytesToManaged(Bytes bytes)
-        {
-            if (bytes == null || bytes.length <= 0 || bytes.b == IntPtr.Zero)
-                return Array.Empty<byte>();
-
-            var raw = new byte[bytes.length];
-            Marshal.Copy(bytes.b, raw, 0, raw.Length);
-            return raw;
-        }
-
-        private static bool TryCreateHaxeBytes(byte[] raw, out Bytes? bytes)
-        {
-            bytes = null;
-
-            try
-            {
-                bytes = Bytes.Class.alloc.Invoke(raw.Length);
-                if (bytes == null)
-                    return false;
-
-                if (raw.Length > 0 && bytes.b != IntPtr.Zero)
-                    Marshal.Copy(raw, 0, bytes.b, raw.Length);
-
-                return true;
-            }
-            catch
-            {
-                bytes = null;
-                return false;
-            }
-        }
-
-        private static int ToInt(object? value)
-        {
-            if (value == null)
-                return 0;
-
-            if (value is int i)
-                return i;
-
-            if (value is bool b)
-                return b ? 1 : 0;
-
-            if (value is IConvertible conv)
-            {
-                try
-                {
-                    return conv.ToInt32(CultureInfo.InvariantCulture);
-                }
-                catch { }
-            }
-
-            return 0;
-        }
-
-        private static ArrayObj EnsureArray(ArrayObj? source)
-        {
-            if (source != null)
-                return source;
-            return (ArrayObj)ArrayUtils.CreateDyn().array;
-        }
-
-        private static ItemMetaManager EnsureItemMeta(User user, ItemMetaManager? meta)
-        {
-            var result = meta ?? user.itemMeta ?? new ItemMetaManager(user);
-            result.itemProgress = EnsureArray(result.itemProgress);
-            result.permanentItems = EnsureArray(result.permanentItems);
-            return result;
-        }
 
         private static LevelGraphSync? CaptureLevelGraph(string levelId, LevelStruct graph)
         {
