+++ ModEntry/ModEntry.ReviveDowned.cs
@@ -1,5 +1,4 @@
 using dc;
-using System.Diagnostics;
 
 
 namespace DeadCellsMultiplayerMod
@@ -12,136 +11,6 @@ namespace DeadCellsMultiplayerMod
         internal static bool IsLocalPlayerDowned()
         {
             return Instance != null && Instance._localFakeDead;
-        }
-
-        /// <summary>
-        /// Boss victory is already host-authoritative through the mob death pipeline.  A downed
-        /// local player is restored as part of that same confirmed event, before reward pickup.
-        /// Duplicate death tombstones are harmless because revive is idempotent.
-        /// </summary>
-        internal static void ReviveLocalPlayerAfterBossVictory()
-        {
-            var instance = Instance;
-            if (instance == null || !instance._localFakeDead)
-                return;
-
-            var net = _net;
-            if (net == null || !net.IsAlive)
-                return;
-
-            instance.ReviveLocalPlayer(net);
-        }
-
-        /// <summary>
-        /// A host-confirmed victory is also the client's presentation barrier: no stale local boss
-        /// death cinematic or spectator target may retain the camera after rewards become available.
-        /// </summary>
-        internal static void RecoverLocalPresentationAfterBossVictory()
-        {
-            var instance = Instance;
-            var net = _net;
-            if (instance == null || net == null || !net.IsAlive || net.IsHost)
-                return;
-
-            instance._clientBossVictoryRecoveryPending = true;
-            instance._clientBossVictoryRecoveryStartedTick = Stopwatch.GetTimestamp();
-            instance._clientBossVictoryRecoveryLevelId = instance.GetCurrentLevelId();
-
-            GameMenu.EnqueueCriticalMainThreadCoalesced("game:boss-victory-presentation", () =>
-            {
-                if (_net == null || !_net.IsAlive || _net.IsHost)
-                    return;
-
-                instance.ApplyClientBossVictoryPresentationRecovery(releaseUnknownCinematic: false);
-            });
-        }
-
-        private void MaintainClientBossVictoryPresentationRecovery()
-        {
-            if (!_clientBossVictoryRecoveryPending)
-                return;
-            if (_netRole != NetRole.Client || _net == null || !_net.IsAlive)
-            {
-                ResetClientBossVictoryPresentationRecovery();
-                return;
-            }
-
-            var currentLevelId = GetCurrentLevelId();
-            if (!string.IsNullOrWhiteSpace(_clientBossVictoryRecoveryLevelId) &&
-                !string.Equals(currentLevelId, _clientBossVictoryRecoveryLevelId, StringComparison.OrdinalIgnoreCase))
-            {
-                ResetClientBossVictoryPresentationRecovery();
-                return;
-            }
-
-            var elapsedSeconds = _clientBossVictoryRecoveryStartedTick > 0
-                ? (Stopwatch.GetTimestamp() - _clientBossVictoryRecoveryStartedTick) / (double)Stopwatch.Frequency
-                : 0.0;
-
-            var hasRetainedCinematic = false;
-            try
-            {
-                var cine = dc.pr.Game.Class.ME?.curCine;
-                hasRetainedCinematic = cine != null;
-            }
-            catch
-            {
-            }
-
-            ApplyClientBossVictoryPresentationRecovery(
-                releaseUnknownCinematic: elapsedSeconds >= ClientBossVictoryUnknownCineReleaseSeconds);
-
-            if (elapsedSeconds >= ClientBossVictoryRecoveryMaxSeconds ||
-                (!hasRetainedCinematic && elapsedSeconds >= ClientBossVictoryNoCineGraceSeconds))
-            {
-                ResetClientBossVictoryPresentationRecovery();
-            }
-        }
-
-        private void ApplyClientBossVictoryPresentationRecovery(bool releaseUnknownCinematic)
-        {
-            try
-            {
-                var game = dc.pr.Game.Class.ME;
-                var cine = game?.curCine;
-                if (cine != null && cine is not DeadBase && cine is not RemoteDownedCorpse)
-                {
-                    var typeName = cine.GetType().Name ?? string.Empty;
-                    var isKnownBossDeath = BossDeathCineTypeNames.Contains(typeName);
-                    var isHeroDeath = typeName.Contains("HeroDeath", StringComparison.OrdinalIgnoreCase);
-                    var isConfirmedBossLevel = IsBossLevel(GetCurrentLevelId());
-                    var isDestroyed = false;
-                    try { isDestroyed = cine.destroyed; } catch { }
-                    if (isKnownBossDeath ||
-                        (isDestroyed && isConfirmedBossLevel && !isHeroDeath) ||
-                        (releaseUnknownCinematic && isConfirmedBossLevel && !isHeroDeath))
-                    {
-                        // Unknown DLC/Boss Rush death cinematics are allowed to finish naturally.
-                        // If one still owns Game.curCine eight seconds after host-confirmed victory,
-                        // it is stale and must not strand only the client behind a letterbox/camera lock.
-                        SuppressRemoteBossDeathCineState(cine);
-                    }
-                }
-            }
-            catch
-            {
-            }
-
-            SuppressRemoteBossDeathCineIfNeeded();
-            _automaticDownedSpectateActive = false;
-            _spectatedRemoteCameraId = 0;
-            _spectatedCameraOrderIndex = 0;
-            try { me?.cancelSkillControlLock(); } catch { }
-            try { me?.unlockControls(); } catch { }
-            EnsureHeroVisibilityAfterRoomChange(me);
-            RequestLocalCameraRefollow("host-boss-victory");
-        }
-
-        private void ResetClientBossVictoryPresentationRecovery()
-        {
-            _clientBossVictoryRecoveryPending = false;
-            _clientBossVictoryRecoveryStartedTick = 0;
-            _clientBossVictoryRecoveryLevelId = null;
         }
 
         internal static bool ShouldAnchorLocalDownedCorpse()
