+++ server/server.NetNode.Cleanup.cs
@@ -12,20 +12,15 @@ public sealed partial class NetNode
             _remotes.Clear();
             _primaryRemoteId = 0;
             _pendingAttacks.Clear();
-            _pendingChatMessages.Clear();
             _pendingMobStates.Clear();
             _pendingMobMoves.Clear();
-            _pendingMobCharges.Clear();
             _pendingMobHits.Clear();
             _pendingMobDies.Clear();
-            _pendingBossVictories.Clear();
             _pendingMobAttacks.Clear();
             _pendingMobDraws.Clear();
             _pendingMobRegistry.Clear();
             _pendingExitReadyStates.Clear();
             _pendingBossCineLevelIds.Clear();
-            _pendingBossIntroEnds.Clear();
-            _pendingBossIntroReadyStates.Clear();
             _pendingBossHeroTeleports.Clear();
             _pendingPlayerDownStates.Clear();
             _pendingPlayerReviveRequests.Clear();
