+++ server/server.NetNode.Consume.cs
@@ -161,24 +161,14 @@ public sealed partial class NetNode
         }
     }
 
-    public bool TryConsumeChatMessages(out List<RemoteChatMessage> messages)
-    {
-        lock (_sync)
-        {
-            return TryConsumePendingListLocked(ref _pendingChatMessages, out messages);
-        }
-    }
-
     public void ClearMobSyncQueues()
     {
         lock (_sync)
         {
             _pendingMobStates.Clear();
             _pendingMobMoves.Clear();
-            _pendingMobCharges.Clear();
             _pendingMobHits.Clear();
             _pendingMobDies.Clear();
-            _pendingBossVictories.Clear();
             _pendingMobAttacks.Clear();
             _pendingMobDraws.Clear();
             _pendingMobRegistry.Clear();
@@ -220,14 +210,6 @@ public sealed partial class NetNode
         }
     }
 
-    public bool TryConsumeMobCharges(out List<MobChargeSnapshot> charges)
-    {
-        lock (_sync)
-        {
-            return TryConsumePendingListLocked(ref _pendingMobCharges, out charges);
-        }
-    }
-
     public bool TryConsumeMobHits(out List<MobHit> hits)
     {
         lock (_sync)
@@ -241,14 +223,6 @@ public sealed partial class NetNode
         lock (_sync)
         {
             return TryConsumePendingListLocked(ref _pendingMobDies, out dies);
-        }
-    }
-
-    public bool TryConsumeBossVictories(out List<BossVictoryState> victories)
-    {
-        lock (_sync)
-        {
-            return TryConsumePendingListLocked(ref _pendingBossVictories, out victories);
         }
     }
 
@@ -302,22 +276,6 @@ public sealed partial class NetNode
         lock (_sync)
         {
             return TryConsumePendingListLocked(ref _pendingBossCineLevelIds, out levelIds);
-        }
-    }
-
-    public bool TryConsumeBossIntroEnds(out List<string> completions)
-    {
-        lock (_sync)
-        {
-            return TryConsumePendingListLocked(ref _pendingBossIntroEnds, out completions);
-        }
-    }
-
-    public bool TryConsumeBossIntroReadyStates(out List<BossIntroReadyState> readyStates)
-    {
-        lock (_sync)
-        {
-            return TryConsumePendingListLocked(ref _pendingBossIntroReadyStates, out readyStates);
         }
     }
